feat: add initial landing page including header, navbar, features, game preview, and footer components.
