+++ src/app/components/pages/indexPage/Index.tsx
@@ -1,0 +1,19 @@
+import Features from './components/Features'
+import GamePreview from './components/GamePreview'
+import Header from './components/Header'
+import Navbar from './components/Navbar'
+import Footer from './components/Footer'
+
+const IndexPage = () => {
+  return (
+    <>
+      <Navbar />
+      <Header />
+      <Features />
+      <GamePreview />
+      <Footer />
+    </>
+  )
+}
+
+export default IndexPage
