+++ src/routes/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-import Index from '@/app/components/pages/Index'
+import Index from '@/app/components/pages/indexPage/Index'
 
 export const Route = createFileRoute('/')({
   component: Index,
